leafsheets-django/requirements.txt
asgiref==3.2.3
asn1crypto==1.1.0
astroid==2.4.2
autopep8==1.5.3
boto3==1.10.35
botocore==1.13.35
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.8
django-address==0.2.1
django-appconf==1.0.3
django-cors-headers==3.2.0
django-environ==0.4.5
django-imagekit==4.0.2
django-jet==1.0.8
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-split-settings==1.0.1
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pilkit==2.0
Pillow==6.2.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.19
pylint==2.5.3
pypandoc==1.4
PyPDF2==1.26.0
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
shortuuid==1.0.1
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
stripe==2.41.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.7
wrapt==1.12.1
